Update ProjectDetailVew tests
