+++ managers/tests/test_api.py
@@ -161,25 +161,3 @@ class CustomProjectsListTests(ProjectTest):
         projects_list = response.context_data["object_list"]
         self.assertTrue(projects_list.ordered)
         self.assertTrue("-members_count" in projects_list.query.order_by)
-
-
-class ProjectDetailTests(ProjectTest):
-    def setUp(self):
-        super().setUp()
-        self.project = ProjectFactory()
-        self.url = reverse("custom-project-detail", args=(self.project.pk,))
-
-    def test_project_detail_view_should_display_project_details_on_get(self):
-        request = APIRequestFactory().get(path=self.url)
-        request.user = self.user
-        response = views.ProjectDetail.as_view()(request, self.project.pk)
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, self.project.name)
-        project = response.data["project"]
-        self.assertEqual(self.project, project)
-
-    def test_project_detail_view_should_return_404_status_code_on_get_if_project_does_not_exist(self):
-        request = APIRequestFactory().get(path=self.url)
-        request.user = self.user
-        response = views.ProjectDetail.as_view()(request, self.project.pk + 1)
-        self.assertEqual(response.status_code, 404)
